+++ src/lib.rs
@@ -197,3 +197,18 @@ pub trait SpongeExt: CryptographicSponge {
     /// Consumes `self` and returns the state.
     fn into_state(self) -> Self::State;
 }
+
+/// The mode structure for duplex sponges
+#[derive(Clone, Debug)]
+pub enum DuplexSpongeMode {
+    /// The sponge is currently absorbing data.
+    Absorbing {
+        /// next position of the state to be XOR-ed when absorbing.
+        next_absorb_index: usize,
+    },
+    /// The sponge is currently squeezing data out.
+    Squeezing {
+        /// next position of the state to be outputted when squeezing.
+        next_squeeze_index: usize,
+    },
+}