+++ src/poseidon/mod.rs
@@ -1,10 +1,9 @@
 use crate::{
     batch_field_cast, squeeze_field_elements_with_sizes_default_impl, Absorb, CryptographicSponge,
-    FieldBasedCryptographicSponge, FieldElementSize, SpongeExt,
+    DuplexSpongeMode, FieldBasedCryptographicSponge, FieldElementSize, SpongeExt,
 };
 use ark_ff::{BigInteger, FpParameters, PrimeField};
 use ark_std::any::TypeId;
-use ark_std::rand::Rng;
 use ark_std::vec;
 use ark_std::vec::Vec;
 
@@ -14,10 +13,32 @@ pub mod constraints;
 #[cfg(test)]
 mod tests;
 
-#[derive(Clone)]
-enum PoseidonSpongeMode {
-    Absorbing { next_absorb_index: usize },
-    Squeezing { next_squeeze_index: usize },
+/// default parameters traits for Poseidon
+pub mod traits;
+pub use traits::*;
+
+mod grain_lfsr;
+
+/// Parameters and RNG used
+#[derive(Clone, Debug)]
+pub struct PoseidonParameters<F: PrimeField> {
+    /// Number of rounds in a full-round operation.
+    pub full_rounds: usize,
+    /// Number of rounds in a partial-round operation.
+    pub partial_rounds: usize,
+    /// Exponent used in S-boxes.
+    pub alpha: u64,
+    /// Additive Round keys. These are added before each MDS matrix application to make it an affine shift.
+    /// They are indexed by `ark[round_num][state_element_index]`
+    pub ark: Vec<Vec<F>>,
+    /// Maximally Distance Separating (MDS) Matrix.
+    pub mds: Vec<Vec<F>>,
+    /// The rate (in terms of number of field elements).
+    /// See [On the Indifferentiability of the Sponge Construction](https://iacr.org/archive/eurocrypt2008/49650180/49650180.pdf)
+    /// for more details on the rate and capacity of a sponge.
+    pub rate: usize,
+    /// The capacity (in terms of number of field elements).
+    pub capacity: usize,
 }
 
 #[derive(Clone)]
@@ -28,28 +49,14 @@ enum PoseidonSpongeMode {
 ///
 /// [cos]: https://eprint.iacr.org/2019/1076
 pub struct PoseidonSponge<F: PrimeField> {
-    // Sponge Parameters
-    /// number of rounds in a full-round operation
-    full_rounds: u32,
-    /// number of rounds in a partial-round operation
-    partial_rounds: u32,
-    /// Exponent used in S-boxes
-    alpha: u64,
-    /// Additive Round keys. These are added before each MDS matrix application to make it an affine shift.
-    /// They are indexed by `ark[round_num][state_element_index]`
-    ark: Vec<Vec<F>>,
-    /// Maximally Distance Separating Matrix.
-    mds: Vec<Vec<F>>,
-    /// the rate (in terms of number of field elements)
-    rate: usize,
-    /// the capacity (in terms of number of field elements)
-    capacity: usize,
+    /// Sponge Parameters
+    pub parameters: PoseidonParameters<F>,
 
     // Sponge State
-    /// current sponge's state (current elements in the permutation block)
-    state: Vec<F>,
-    /// current mode (whether its absorbing or squeezing)
-    mode: PoseidonSpongeMode,
+    /// Current sponge's state (current elements in the permutation block)
+    pub state: Vec<F>,
+    /// Current mode (whether its absorbing or squeezing)
+    pub mode: DuplexSpongeMode,
 }
 
 impl<F: PrimeField> PoseidonSponge<F> {
@@ -57,18 +64,18 @@ impl<F: PrimeField> PoseidonSponge<F> {
         // Full rounds apply the S Box (x^alpha) to every element of state
         if is_full_round {
             for elem in state {
-                *elem = elem.pow(&[self.alpha]);
+                *elem = elem.pow(&[self.parameters.alpha]);
             }
         }
-        // Partial rounds apply the S Box (x^alpha) to just the final element of state
+        // Partial rounds apply the S Box (x^alpha) to just the first element of state
         else {
-            state[state.len() - 1] = state[state.len() - 1].pow(&[self.alpha]);
+            state[0] = state[0].pow(&[self.parameters.alpha]);
         }
     }
 
     fn apply_ark(&self, state: &mut [F], round_number: usize) {
         for (i, state_elem) in state.iter_mut().enumerate() {
-            state_elem.add_assign(&self.ark[round_number][i]);
+            state_elem.add_assign(&self.parameters.ark[round_number][i]);
         }
     }
 
@@ -77,7 +84,7 @@ impl<F: PrimeField> PoseidonSponge<F> {
         for i in 0..state.len() {
             let mut cur = F::zero();
             for (j, state_elem) in state.iter().enumerate() {
-                let term = state_elem.mul(&self.mds[i][j]);
+                let term = state_elem.mul(&self.parameters.mds[i][j]);
                 cur.add_assign(&term);
             }
             new_state.push(cur);
@@ -86,24 +93,24 @@ impl<F: PrimeField> PoseidonSponge<F> {
     }
 
     fn permute(&mut self) {
-        let full_rounds_over_2 = self.full_rounds / 2;
+        let full_rounds_over_2 = self.parameters.full_rounds / 2;
         let mut state = self.state.clone();
         for i in 0..full_rounds_over_2 {
-            self.apply_ark(&mut state, i as usize);
+            self.apply_ark(&mut state, i);
             self.apply_s_box(&mut state, true);
             self.apply_mds(&mut state);
         }
 
-        for i in full_rounds_over_2..(full_rounds_over_2 + self.partial_rounds) {
-            self.apply_ark(&mut state, i as usize);
+        for i in full_rounds_over_2..(full_rounds_over_2 + self.parameters.partial_rounds) {
+            self.apply_ark(&mut state, i);
             self.apply_s_box(&mut state, false);
             self.apply_mds(&mut state);
         }
 
-        for i in
-            (full_rounds_over_2 + self.partial_rounds)..(self.partial_rounds + self.full_rounds)
+        for i in (full_rounds_over_2 + self.parameters.partial_rounds)
+            ..(self.parameters.partial_rounds + self.parameters.full_rounds)
         {
-            self.apply_ark(&mut state, i as usize);
+            self.apply_ark(&mut state, i);
             self.apply_s_box(&mut state, true);
             self.apply_mds(&mut state);
         }
@@ -116,24 +123,24 @@ impl<F: PrimeField> PoseidonSponge<F> {
 
         loop {
             // if we can finish in this call
-            if rate_start_index + remaining_elements.len() <= self.rate {
+            if rate_start_index + remaining_elements.len() <= self.parameters.rate {
                 for (i, element) in remaining_elements.iter().enumerate() {
-                    self.state[i + rate_start_index] += element;
+                    self.state[self.parameters.capacity + i + rate_start_index] += element;
                 }
-                self.mode = PoseidonSpongeMode::Absorbing {
+                self.mode = DuplexSpongeMode::Absorbing {
                     next_absorb_index: rate_start_index + remaining_elements.len(),
                 };
 
                 return;
             }
             // otherwise absorb (rate - rate_start_index) elements
-            let num_elements_absorbed = self.rate - rate_start_index;
+            let num_elements_absorbed = self.parameters.rate - rate_start_index;
             for (i, element) in remaining_elements
                 .iter()
                 .enumerate()
                 .take(num_elements_absorbed)
             {
-                self.state[i + rate_start_index] += element;
+                self.state[self.parameters.capacity + i + rate_start_index] += element;
             }
             self.permute();
             // the input elements got truncated by num elements absorbed
@@ -147,23 +154,25 @@ impl<F: PrimeField> PoseidonSponge<F> {
         let mut output_remaining = output;
         loop {
             // if we can finish in this call
-            if rate_start_index + output_remaining.len() <= self.rate {
+            if rate_start_index + output_remaining.len() <= self.parameters.rate {
                 output_remaining.clone_from_slice(
-                    &self.state[rate_start_index..(output_remaining.len() + rate_start_index)],
+                    &self.state[self.parameters.capacity + rate_start_index
+                        ..(self.parameters.capacity + output_remaining.len() + rate_start_index)],
                 );
-                self.mode = PoseidonSpongeMode::Squeezing {
+                self.mode = DuplexSpongeMode::Squeezing {
                     next_squeeze_index: rate_start_index + output_remaining.len(),
                 };
                 return;
             }
             // otherwise squeeze (rate - rate_start_index) elements
-            let num_elements_squeezed = self.rate - rate_start_index;
+            let num_elements_squeezed = self.parameters.rate - rate_start_index;
             output_remaining[..num_elements_squeezed].clone_from_slice(
-                &self.state[rate_start_index..(num_elements_squeezed + rate_start_index)],
+                &self.state[self.parameters.capacity + rate_start_index
+                    ..(self.parameters.capacity + num_elements_squeezed + rate_start_index)],
             );
 
             // Unless we are done with squeezing in this call, permute.
-            if output_remaining.len() != self.rate {
+            if output_remaining.len() != self.parameters.rate {
                 self.permute();
             }
             // Repeat with updated output slices
@@ -173,29 +182,24 @@ impl<F: PrimeField> PoseidonSponge<F> {
     }
 }
 
-/// Parameters and RNG used
-#[derive(Clone, Debug)]
-pub struct PoseidonParameters<F: PrimeField> {
-    full_rounds: u32,
-    partial_rounds: u32,
-    alpha: u64,
-    mds: Vec<Vec<F>>,
-    ark: Vec<Vec<F>>,
-}
-
 impl<F: PrimeField> PoseidonParameters<F> {
     /// Initialize the parameter for Poseidon Sponge.
     pub fn new(
-        full_rounds: u32,
-        partial_rounds: u32,
+        full_rounds: usize,
+        partial_rounds: usize,
         alpha: u64,
         mds: Vec<Vec<F>>,
         ark: Vec<Vec<F>>,
+        rate: usize,
+        capacity: usize,
     ) -> Self {
-        // shape check
-        assert_eq!(ark.len() as u32, full_rounds + partial_rounds);
+        assert_eq!(ark.len(), full_rounds + partial_rounds);
         for item in &ark {
-            assert_eq!(item.len(), 3);
+            assert_eq!(item.len(), rate + capacity);
+        }
+        assert_eq!(mds.len(), rate + capacity);
+        for item in &mds {
+            assert_eq!(item.len(), rate + capacity);
         }
         Self {
             full_rounds,
@@ -203,56 +207,24 @@ impl<F: PrimeField> PoseidonParameters<F> {
             alpha,
             mds,
             ark,
+            rate,
+            capacity,
         }
-    }
-
-    /// Return a random round constant.
-    pub fn random_ark<R: Rng>(full_rounds: u32, partial_rounds: u32, rng: &mut R) -> Vec<Vec<F>> {
-        let mut ark = Vec::new();
-
-        for _ in 0..full_rounds + partial_rounds {
-            let mut res = Vec::new();
-
-            for _ in 0..3 {
-                res.push(F::rand(rng));
-            }
-            ark.push(res);
-        }
-
-        ark
     }
 }
 
 impl<F: PrimeField> CryptographicSponge for PoseidonSponge<F> {
     type Parameters = PoseidonParameters<F>;
 
-    fn new(params: &Self::Parameters) -> Self {
-        // Requires F to be Alt_Bn128Fr
-        let full_rounds = params.full_rounds;
-        let partial_rounds = params.partial_rounds;
-        let alpha = params.alpha;
-
-        let mds = params.mds.clone();
-
-        let ark = params.ark.to_vec();
-
-        let rate = 2;
-        let capacity = 1;
-        let state = vec![F::zero(); rate + capacity];
-        let mode = PoseidonSpongeMode::Absorbing {
+    fn new(parameters: &Self::Parameters) -> Self {
+        let state = vec![F::zero(); parameters.rate + parameters.capacity];
+        let mode = DuplexSpongeMode::Absorbing {
             next_absorb_index: 0,
         };
 
         Self {
-            full_rounds,
-            partial_rounds,
-            alpha,
-            ark,
-            mds,
-
+            parameters: parameters.clone(),
             state,
-            rate,
-            capacity,
             mode,
         }
     }
@@ -264,15 +236,15 @@ impl<F: PrimeField> CryptographicSponge for PoseidonSponge<F> {
         }
 
         match self.mode {
-            PoseidonSpongeMode::Absorbing { next_absorb_index } => {
+            DuplexSpongeMode::Absorbing { next_absorb_index } => {
                 let mut absorb_index = next_absorb_index;
-                if absorb_index == self.rate {
+                if absorb_index == self.parameters.rate {
                     self.permute();
                     absorb_index = 0;
                 }
                 self.absorb_internal(absorb_index, elems.as_slice());
             }
-            PoseidonSpongeMode::Squeezing {
+            DuplexSpongeMode::Squeezing {
                 next_squeeze_index: _,
             } => {
                 self.permute();
@@ -349,15 +321,15 @@ impl<F: PrimeField> FieldBasedCryptographicSponge<F> for PoseidonSponge<F> {
     fn squeeze_native_field_elements(&mut self, num_elements: usize) -> Vec<F> {
         let mut squeezed_elems = vec![F::zero(); num_elements];
         match self.mode {
-            PoseidonSpongeMode::Absorbing {
+            DuplexSpongeMode::Absorbing {
                 next_absorb_index: _,
             } => {
                 self.permute();
                 self.squeeze_internal(0, &mut squeezed_elems);
             }
-            PoseidonSpongeMode::Squeezing { next_squeeze_index } => {
+            DuplexSpongeMode::Squeezing { next_squeeze_index } => {
                 let mut squeeze_index = next_squeeze_index;
-                if squeeze_index == self.rate {
+                if squeeze_index == self.parameters.rate {
                     self.permute();
                     squeeze_index = 0;
                 }
@@ -373,7 +345,7 @@ impl<F: PrimeField> FieldBasedCryptographicSponge<F> for PoseidonSponge<F> {
 /// Stores the state of a Poseidon Sponge. Does not store any parameter.
 pub struct PoseidonSpongeState<F: PrimeField> {
     state: Vec<F>,
-    mode: PoseidonSpongeMode,
+    mode: DuplexSpongeMode,
 }
 
 impl<CF: PrimeField> SpongeExt for PoseidonSponge<CF> {
@@ -391,5 +363,97 @@ impl<CF: PrimeField> SpongeExt for PoseidonSponge<CF> {
             state: self.state,
             mode: self.mode,
         }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::poseidon::{
+        PoseidonDefaultParameters, PoseidonDefaultParametersEntry, PoseidonDefaultParametersField,
+    };
+    use crate::{poseidon::PoseidonSponge, CryptographicSponge, FieldBasedCryptographicSponge};
+    use ark_ff::{field_new, BigInteger256, FftParameters, Fp256, Fp256Parameters, FpParameters};
+    use ark_test_curves::bls12_381::FrParameters;
+
+    pub struct TestFrParameters;
+
+    impl Fp256Parameters for TestFrParameters {}
+    impl FftParameters for TestFrParameters {
+        type BigInt = <FrParameters as FftParameters>::BigInt;
+        const TWO_ADICITY: u32 = FrParameters::TWO_ADICITY;
+        const TWO_ADIC_ROOT_OF_UNITY: Self::BigInt = FrParameters::TWO_ADIC_ROOT_OF_UNITY;
+    }
+
+    // This TestFrParameters is the same as the BLS12-381's Fr.
+    // MODULUS = 52435875175126190479447740508185965837690552500527637822603658699938581184513
+    impl FpParameters for TestFrParameters {
+        const MODULUS: BigInteger256 = FrParameters::MODULUS;
+        const MODULUS_BITS: u32 = FrParameters::MODULUS_BITS;
+        const CAPACITY: u32 = FrParameters::CAPACITY;
+        const REPR_SHAVE_BITS: u32 = FrParameters::REPR_SHAVE_BITS;
+        const R: BigInteger256 = FrParameters::R;
+        const R2: BigInteger256 = FrParameters::R2;
+        const INV: u64 = FrParameters::INV;
+        const GENERATOR: BigInteger256 = FrParameters::GENERATOR;
+        const MODULUS_MINUS_ONE_DIV_TWO: BigInteger256 = FrParameters::MODULUS_MINUS_ONE_DIV_TWO;
+        const T: BigInteger256 = FrParameters::T;
+        const T_MINUS_ONE_DIV_TWO: BigInteger256 = FrParameters::T_MINUS_ONE_DIV_TWO;
+    }
+
+    impl PoseidonDefaultParameters for TestFrParameters {
+        const PARAMS_OPT_FOR_CONSTRAINTS: [PoseidonDefaultParametersEntry; 7] = [
+            PoseidonDefaultParametersEntry::new(2, 17, 8, 31, 0),
+            PoseidonDefaultParametersEntry::new(3, 5, 8, 56, 0),
+            PoseidonDefaultParametersEntry::new(4, 5, 8, 56, 0),
+            PoseidonDefaultParametersEntry::new(5, 5, 8, 57, 0),
+            PoseidonDefaultParametersEntry::new(6, 5, 8, 57, 0),
+            PoseidonDefaultParametersEntry::new(7, 5, 8, 57, 0),
+            PoseidonDefaultParametersEntry::new(8, 5, 8, 57, 0),
+        ];
+        const PARAMS_OPT_FOR_WEIGHTS: [PoseidonDefaultParametersEntry; 7] = [
+            PoseidonDefaultParametersEntry::new(2, 257, 8, 13, 0),
+            PoseidonDefaultParametersEntry::new(3, 257, 8, 13, 0),
+            PoseidonDefaultParametersEntry::new(4, 257, 8, 13, 0),
+            PoseidonDefaultParametersEntry::new(5, 257, 8, 13, 0),
+            PoseidonDefaultParametersEntry::new(6, 257, 8, 13, 0),
+            PoseidonDefaultParametersEntry::new(7, 257, 8, 13, 0),
+            PoseidonDefaultParametersEntry::new(8, 257, 8, 13, 0),
+        ];
+    }
+
+    pub type TestFr = Fp256<TestFrParameters>;
+
+    #[test]
+    fn test_poseidon_sponge_consistency() {
+        let sponge_param = TestFr::get_default_poseidon_parameters(2, false).unwrap();
+
+        let mut sponge = PoseidonSponge::<TestFr>::new(&sponge_param);
+        sponge.absorb(&vec![
+            TestFr::from(0u8),
+            TestFr::from(1u8),
+            TestFr::from(2u8),
+        ]);
+        let res = sponge.squeeze_native_field_elements(3);
+        assert_eq!(
+            res[0],
+            field_new!(
+                TestFr,
+                "40442793463571304028337753002242186710310163897048962278675457993207843616876"
+            )
+        );
+        assert_eq!(
+            res[1],
+            field_new!(
+                TestFr,
+                "2664374461699898000291153145224099287711224021716202960480903840045233645301"
+            )
+        );
+        assert_eq!(
+            res[2],
+            field_new!(
+                TestFr,
+                "50191078828066923662070228256530692951801504043422844038937334196346054068797"
+            )
+        );
     }
 }