+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Breaking changes
 
-- [\#21](https://github.com/arkworks-rs/sponge/pull/21) Fix Poseidon `random_ark` being too short.
+- [\#22](https://github.com/arkworks-rs/sponge/pull/22) Clean up the Poseidon parameter and sponge structures.
 
 ### Features
+
+- [\#22](https://github.com/arkworks-rs/sponge/pull/22) Add traits and derivations for default Poseidon parameters.
 
 ### Improvements